Rounded out User Container guide
Filled in most of the gaps and restructured the guide to walk a user
through the concepts of creating profiles and a container workflow.

This does not touch on much of anything once the container is up and
running. That will come later once the Container Exit Strategy has been
revamped.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>