fix outdated links (#251)
* containers: update examples and links

Problem: the user container docs refer to v1alpha1, but that version
is no longer supported and the links in the docs break because the
files have been removed from the repo.

Fix the links and examples.

* data movement: update links

Problem: the data movement docs refer to v1alpha1, but that version
is no longer supported and the links in the docs break because the
files have been removed from the repo.

Fix the links.